Move the TEOS-10 tracer conversion to polaris.ocean.eos
Per review feedback, convert_tracers_to_mpas_ocean() is an EOS/tracer-
convention conversion, not something p-star specific, so move it from
polaris.ocean.vertical.pstar_state to polaris/ocean/eos/teos10.py and
export it from polaris.ocean.eos where developers will look for it.
